shell: refactor shunit2 subsystem
