+++ src/python/pants/backend/shell/register.py
@@ -1,8 +1,9 @@
 # Copyright 2021 Pants project contributors (see CONTRIBUTORS.md).
 # Licensed under the Apache License, Version 2.0 (see LICENSE).
 
-from pants.backend.shell import dependency_inference, shunit2, shunit2_test_runner
+from pants.backend.shell import dependency_inference, shunit2_test_runner
 from pants.backend.shell.goals import tailor, test
+from pants.backend.shell.subsystems import shunit2
 from pants.backend.shell.target_types import (
     ShellCommandRunTarget,
     ShellCommandTarget,
